tests unitaires plats + debut carte

+++ organisation_carte/models/plat.rb
@@ -6,7 +6,7 @@ module Sample
       field :nom, :type => String
       field :description, :type => String
       field :type, :type => String
-      field :prix, :type => Integer
+      field :prix, :type => Float
       field :quantite, :type => Integer
 
     end
